+++ plugin-enclave-gradle/src/main/kotlin/com/r3/conclave/plugin/enclave/gradle/SignEnclave.kt
@@ -1,6 +1,5 @@
 package com.r3.conclave.plugin.enclave.gradle
 
-import com.r3.conclave.plugin.enclave.gradle.extension.EnclaveExtension
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.tasks.InputFile
